sync update function in admin view reward, auto add redeemedCount in the roomDB after user redeem the reward

+++ app/src/main/java/com/example/kleine/viewmodel/admin/AdminViewRewardViewModel.kt
@@ -12,6 +12,7 @@ import androidx.lifecycle.viewModelScope
 import com.example.kleine.database.Reward
 import com.example.kleine.database.RewardDao
 import com.example.kleine.resource.NetworkReceiver
+import com.google.firebase.firestore.CollectionReference
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.storage.FirebaseStorage
 import kotlinx.coroutines.launch
@@ -55,6 +56,14 @@ class AdminViewRewardViewModel(private val appContext: Context, private val rewa
     }
 
     private fun syncLocalDataToFirestore() {
+        viewModelScope.launch {
+            syncNewRewardsToFirestore()
+            updateFirestoreForModifiedRewards()
+        }
+    }
+
+
+    private fun syncNewRewardsToFirestore() {
         viewModelScope.launch {
             val unsyncedRewards = rewardDao.getUnsyncedRewards(1) // Assuming isAdded = 1 denotes unsynced data
             val rewardsCollection = FirebaseFirestore.getInstance().collection("Rewards")
@@ -106,6 +115,69 @@ class AdminViewRewardViewModel(private val appContext: Context, private val rewa
         }
     }
 
+    private fun updateFirestoreForModifiedRewards() {
+        viewModelScope.launch {
+            val modifiedRewards = rewardDao.getModifiedRewards(1) // Assuming isUpdated = 1 denotes modified data
+            val rewardsCollection = FirebaseFirestore.getInstance().collection("Rewards")
+
+            for (reward in modifiedRewards) {
+                if (reward.imageUrl == "changed" && reward.imageBytes != null) {
+                    // Update the image on Firestore
+                    val imageRef = firebaseStorage.reference.child("rewards/${System.currentTimeMillis()}.jpg")
+
+                    imageRef.putBytes(reward.imageBytes!!).addOnSuccessListener {
+                        imageRef.downloadUrl.addOnSuccessListener { uri ->
+                            val newImageUrl = uri.toString()
+
+                            // Create a new data map to represent the reward for Firestore
+                            val firestoreReward = hashMapOf(
+                                "rewardName" to reward.rewardName,
+                                "imageUrl" to newImageUrl,
+                                "redeemLimit" to reward.redeemLimit,
+                                "redeemedCount" to reward.redeemedCount,
+                                "rewardDescription" to reward.rewardDescription,
+                                "rewardPoints" to reward.rewardPoints,
+                            )
+
+                            updateFirestoreAndResetImageUrl(rewardsCollection, reward, firestoreReward as HashMap<String, Any?>)
+                        }
+                    }.addOnFailureListener { exception ->
+                        val errorMsg = "Error uploading image: ${exception.message}"
+                        Log.e("AdminViewRewardVM", errorMsg)
+                    }
+                } else {
+                    // Don't update the image on Firestore, but other data can be updated
+                    val firestoreReward = hashMapOf(
+                        "rewardName" to reward.rewardName,
+                        "redeemLimit" to reward.redeemLimit,
+                        "redeemedCount" to reward.redeemedCount,
+                        "rewardDescription" to reward.rewardDescription,
+                        "rewardPoints" to reward.rewardPoints,
+                    )
+
+                    updateFirestoreAndResetImageUrl(rewardsCollection, reward, firestoreReward as HashMap<String, Any?>)
+                }
+            }
+        }
+    }
+
+    private fun updateFirestoreAndResetImageUrl(rewardsCollection: CollectionReference, reward: Reward, firestoreReward: HashMap<String, Any?>) {
+        rewardsCollection.whereEqualTo("rewardName", reward.rewardName)
+            .get()
+            .addOnSuccessListener { querySnapshot ->
+                if (!querySnapshot.isEmpty) {
+                    val document = querySnapshot.documents[0]
+                    document.reference.update(firestoreReward as Map<String, Any>).addOnSuccessListener {
+                        // Once updated successfully, mark the reward as not modified and imageUrl to null in Room
+                        reward.isUpdated = 0
+                        reward.imageUrl = null
+                        viewModelScope.launch {
+                            rewardDao.update(reward)
+                        }
+                    }
+                }
+            }
+    }
 
     private fun fetchRewardsFromFirestore() {
         val rewardsCollection = FirebaseFirestore.getInstance().collection("Rewards")
@@ -134,37 +206,51 @@ class AdminViewRewardViewModel(private val appContext: Context, private val rewa
     }
 
     fun deleteReward(rewardName: String) {
-        val rewardsCollection = FirebaseFirestore.getInstance().collection("Rewards")
+        if(isNetworkAvailable){
+            val rewardsCollection = FirebaseFirestore.getInstance().collection("Rewards")
 
-        // Query to find the document that has the rewardName
-        rewardsCollection.whereEqualTo("rewardName", rewardName)
-            .get()
-            .addOnSuccessListener { querySnapshot ->
-                // Check if a document with that rewardName was found
-                if (!querySnapshot.isEmpty) {
-                    // Get the first (and should be the only) document that matches the query
-                    val document = querySnapshot.documents[0]
+            // Query to find the document that has the rewardName
+            rewardsCollection.whereEqualTo("rewardName", rewardName)
+                .get()
+                .addOnSuccessListener { querySnapshot ->
+                    // Check if a document with that rewardName was found
+                    if (!querySnapshot.isEmpty) {
+                        // Get the first (and should be the only) document that matches the query
+                        val document = querySnapshot.documents[0]
 
-                    // Delete the found document
-                    document.reference
-                        .delete()
-                        .addOnSuccessListener {
-                            // Log success
-                            deleteResult.value = true
-                        }
-                        .addOnFailureListener { e ->
-                            // Handle failure
-                            deleteResult.value = false
-                        }
-                } else {
-                    // No document with that rewardName found
+                        // Delete the found document
+                        document.reference
+                            .delete()
+                            .addOnSuccessListener {
+                                // Log success
+                                deleteResult.value = true
+
+                                // Mark reward as deleted in RoomDB
+                                viewModelScope.launch {
+                                    val existingReward = rewardDao.getRewardByName(rewardName)
+                                    if(existingReward != null) {
+                                        existingReward.isDeleted = 1
+                                        rewardDao.update(existingReward)
+                                    }
+                                }
+                            }
+                            .addOnFailureListener { e ->
+                                // Handle failure
+                                deleteResult.value = false
+                            }
+                    } else {
+                        // No document with that rewardName found
+                        deleteResult.value = false
+                    }
+                }
+                .addOnFailureListener { e ->
+                    // Handle failure
                     deleteResult.value = false
                 }
-            }
-            .addOnFailureListener { e ->
-                // Handle failure
-                deleteResult.value = false
-            }
+        } else {
+            deleteResult.value = false
+        }
+
     }
